Success refactor, removed 4 files, need to do css
