--- src/components/work/WorkItems/ItemStyles.module.css
@@ -1,118 +1,0 @@
-.workItem {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: flex-start;
-}
-
-.textBox {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  align-content: center;
-  width: 300px;
-  height: auto;
-}
-.title {
-  font-size: var(--mobileHeading);
-  font-weight: bold;
-  color: var(--light-black);
-  text-align: center;
-  margin-bottom: 0.5rem;
-}
-.description {
-  font-size: 1rem;
-  color: var(--black);
-  margin-bottom: 0.5rem;
-}
-.link {
-  height: auto;
-  width: auto;
-  padding-top: 1rem;
-  font-size: 1.2rem;
-  text-align: center;
-  margin-bottom: 0.5rem;
-  text-decoration: none;
-}
-
-.svg:hover {
-  cursor: pointer;
-}
-
-@media screen and (min-width: 768px) {
-  .workItem {
-    padding-top: 1rem;
-    /* width: calc(auto+2rem); */
-  }
-  .textBox {
-    width: 100%;
-    height: auto;
-    margin-left: 0.5rem;
-  }
-  .title {
-    font-size: 2.5rem;
-    font-weight: bold;
-
-    text-align: center;
-    margin-bottom: 0.5rem;
-  }
-  .description {
-    /* font-size: 1.25rem; */
-    text-align: center;
-    margin-bottom: 0.5rem;
-  }
-  .link {
-    height: 100%;
-    width: 100%;
-    padding-top: 1rem;
-    font-size: 1.2rem;
-    text-align: center;
-    margin-bottom: 0.5rem;
-    text-decoration: none;
-  }
-  .svg:hover {
-    cursor: pointer;
-  }
-
-  .Button {
-    min-height: 4rem;
-    border-radius: 5px;
-    color: var(--cyan);
-  }
-}
-
-@media screen and (min-width: 1024px) {
-  .workItem {
-  }
-  .textBox {
-    width: 100%;
-    height: auto;
-    margin-left: 0.5rem;
-  }
-  .title {
-    font-size: 1.75rem;
-    font-weight: bold;
-
-    text-align: center;
-    margin-bottom: 0.5rem;
-  }
-  .description {
-    font-size: 1.15rem;
-    text-align: center;
-    margin-bottom: 0.5rem;
-  }
-  .link {
-    /* display: none; */
-    font-weight: bold;
-    height: 100%;
-    width: 100%;
-    /* padding-top: 1rem; */
-    font-size: 1.2rem;
-    text-align: center;
-    margin-bottom: 0.5rem;
-    text-decoration: none;
-  }
-  .svg:hover {
-    cursor: pointer;
-  }
-}
